+++ Source/ChaosEngine.cpp
@@ -303,7 +303,7 @@ void ChaosEngine::DrawModule()
    ofPopStyle();
 }
 
-void ChaosEngine::OnClicked(int x, int y, bool right)
+void ChaosEngine::OnClicked(float x, float y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
